+++ bolt/internal/internal.go
@@ -171,17 +171,14 @@ func MarshalDashboard(d chronograf.Dashboard) ([]byte, error) {
 				r.Upper, r.Lower = q.Range.Upper, q.Range.Lower
 			}
 			queries[j] = &Query{
-				Command:  q.Command,
-				DB:       q.DB,
-				RP:       q.RP,
-				GroupBys: q.GroupBys,
-				Wheres:   q.Wheres,
-				Label:    q.Label,
-				Range:    r,
+				Command: q.Command,
+				Label:   q.Label,
+				Range:   r,
 			}
 		}
 
 		cells[i] = &DashboardCell{
+			ID:      c.ID,
 			X:       c.X,
 			Y:       c.Y,
 			W:       c.W,
@@ -208,15 +205,11 @@ func UnmarshalDashboard(data []byte, d *chronograf.Dashboard) error {
 
 	cells := make([]chronograf.DashboardCell, len(pb.Cells))
 	for i, c := range pb.Cells {
-		queries := make([]chronograf.Query, len(c.Queries))
+		queries := make([]chronograf.DashboardQuery, len(c.Queries))
 		for j, q := range c.Queries {
-			queries[j] = chronograf.Query{
-				Command:  q.Command,
-				DB:       q.DB,
-				RP:       q.RP,
-				GroupBys: q.GroupBys,
-				Wheres:   q.Wheres,
-				Label:    q.Label,
+			queries[j] = chronograf.DashboardQuery{
+				Command: q.Command,
+				Label:   q.Label,
 			}
 			if q.Range.Upper != q.Range.Lower {
 				queries[j].Range = &chronograf.Range{
@@ -227,6 +220,7 @@ func UnmarshalDashboard(data []byte, d *chronograf.Dashboard) error {
 		}
 
 		cells[i] = chronograf.DashboardCell{
+			ID:      c.ID,
 			X:       c.X,
 			Y:       c.Y,
 			W:       c.W,