+++ ui/src/dashboards/components/Dashboard.js
@@ -12,6 +12,7 @@ const Dashboard = ({
   onEditCell,
   onRenameCell,
   onUpdateCell,
+  onSummonOverlayTechnologies,
   source,
   autoRefresh,
   timeRange,
@@ -24,20 +25,26 @@ const Dashboard = ({
     <div className={classnames({'page-contents': true, 'presentation-mode': inPresentationMode})}>
       <div className={classnames('container-fluid full-width dashboard', {'dashboard-edit': isEditMode})}>
         {isEditMode ? <Visualizations/> : null}
-        {Dashboard.renderDashboard(dashboard, autoRefresh, timeRange, source, onPositionChange, onEditCell, onRenameCell, onUpdateCell)}
+        {Dashboard.renderDashboard(dashboard, autoRefresh, timeRange, source, onPositionChange, onEditCell, onRenameCell, onUpdateCell, onSummonOverlayTechnologies)}
       </div>
     </div>
   )
 }
 
-Dashboard.renderDashboard = (dashboard, autoRefresh, timeRange, source, onPositionChange, onEditCell, onRenameCell, onUpdateCell) => {
+Dashboard.renderDashboard = (dashboard, autoRefresh, timeRange, source, onPositionChange, onEditCell, onRenameCell, onUpdateCell, onSummonOverlayTechnologies) => {
   const cells = dashboard.cells.map((cell, i) => {
     i = `${i}`
     const dashboardCell = {...cell, i}
-    dashboardCell.queries.forEach((q) => {
-      q.text = q.query;
-      q.database = q.db;
-    });
+    dashboardCell.queries = dashboardCell.queries.map(({label, query, queryConfig, db}) =>
+      ({
+        label,
+        query,
+        queryConfig,
+        db,
+        database: db,
+        text: query,
+      })
+    )
     return dashboardCell;
   })
 
@@ -51,6 +58,7 @@ Dashboard.renderDashboard = (dashboard, autoRefresh, timeRange, source, onPositi
       onEditCell={onEditCell}
       onRenameCell={onRenameCell}
       onUpdateCell={onUpdateCell}
+      onSummonOverlayTechnologies={onSummonOverlayTechnologies}
     />
   )
 }
@@ -71,6 +79,7 @@ Dashboard.propTypes = {
   onEditCell: func,
   onRenameCell: func,
   onUpdateCell: func,
+  onSummonOverlayTechnologies: func,
   source: shape({
     links: shape({
       proxy: string,