+++ ui/src/dashboards/containers/DashboardPage.js
@@ -3,13 +3,14 @@ import {Link} from 'react-router'
 import {connect} from 'react-redux'
 import {bindActionCreators} from 'redux'
 
+import CellEditorOverlay from 'src/dashboards/components/CellEditorOverlay'
 import Header from 'src/dashboards/components/DashboardHeader'
 import EditHeader from 'src/dashboards/components/DashboardHeaderEdit'
 import Dashboard from 'src/dashboards/components/Dashboard'
-import timeRanges from 'hson!../../shared/data/timeRanges.hson'
 
 import * as dashboardActionCreators from 'src/dashboards/actions'
 
+import {setAutoRefresh} from 'shared/actions/app'
 import {presentationButtonDispatcher} from 'shared/dispatchers'
 
 const {
@@ -53,11 +54,31 @@ const DashboardPage = React.createClass({
       id: number.isRequired,
       cells: arrayOf(shape({})).isRequired,
     }).isRequired,
+    handleChooseAutoRefresh: func.isRequired,
     autoRefresh: number.isRequired,
     timeRange: shape({}).isRequired,
     inPresentationMode: bool.isRequired,
     isEditMode: bool.isRequired,
     handleClickPresentationButton: func,
+  },
+
+  childContextTypes: {
+    source: shape({
+      links: shape({
+        proxy: string.isRequired,
+        self: string.isRequired,
+      }).isRequired,
+    }).isRequired,
+  },
+
+  getChildContext() {
+    return {source: this.props.source};
+  },
+
+  getInitialState() {
+    return {
+      selectedCell: null,
+    }
   },
 
   componentDidMount() {
@@ -85,9 +106,21 @@ const DashboardPage = React.createClass({
     setEditMode(nextPathname.includes('/edit'))
   },
 
+  handleDismissOverlay() {
+    this.setState({selectedCell: null})
+  },
+
+  handleSaveEditedCell(newCell) {
+    this.props.dashboardActions.updateDashboardCell(newCell)
+    .then(this.handleDismissOverlay)
+  },
+
+  handleSummonOverlayTechnologies(cell) {
+    this.setState({selectedCell: cell})
+  },
+
   handleChooseTimeRange({lower}) {
-    const timeRange = timeRanges.find((range) => range.queryValue === lower);
-    this.props.dashboardActions.setTimeRange(timeRange)
+    this.props.dashboardActions.setTimeRange({lower, upper: null})
   },
 
   handleUpdatePosition(cells) {
@@ -124,17 +157,34 @@ const DashboardPage = React.createClass({
       isEditMode,
       handleClickPresentationButton,
       source,
+      handleChooseAutoRefresh,
       autoRefresh,
       timeRange,
     } = this.props
 
+    const {
+      selectedCell,
+    } = this.state
+
     return (
       <div className="page">
+        {
+          selectedCell && selectedCell.queries.length ?
+            <CellEditorOverlay
+              cell={selectedCell}
+              autoRefresh={autoRefresh}
+              timeRange={timeRange}
+              onCancel={this.handleDismissOverlay}
+              onSave={this.handleSaveEditedCell}
+            /> :
+            null
+        }
         {
           isEditMode ?
             <EditHeader dashboard={dashboard} onSave={() => {}} /> :
             <Header
               buttonText={dashboard ? dashboard.name : ''}
+              handleChooseAutoRefresh={handleChooseAutoRefresh}
               autoRefresh={autoRefresh}
               timeRange={timeRange}
               handleChooseTimeRange={this.handleChooseTimeRange}
@@ -166,6 +216,7 @@ const DashboardPage = React.createClass({
           onEditCell={this.handleEditCell}
           onRenameCell={this.handleChangeCellName}
           onUpdateCell={this.handleUpdateCell}
+          onSummonOverlayTechnologies={this.handleSummonOverlayTechnologies}
         />
       </div>
     );
@@ -197,6 +248,7 @@ const mapStateToProps = (state) => {
 }
 
 const mapDispatchToProps = (dispatch) => ({
+  handleChooseAutoRefresh: bindActionCreators(setAutoRefresh, dispatch),
   handleClickPresentationButton: presentationButtonDispatcher(dispatch),
   dashboardActions: bindActionCreators(dashboardActionCreators, dispatch),
 })