+++ ui/src/hosts/containers/HostPage.js
@@ -88,7 +88,7 @@ export const HostPage = React.createClass({
   },
 
   handleChooseTimeRange({lower}) {
-    const timeRange = timeRanges.find((range) => range.queryValue === lower);
+    const timeRange = timeRanges.find((range) => range.lower === lower);
     this.setState({timeRange});
   },