+++ ui/src/shared/components/LayoutRenderer.js
@@ -4,6 +4,9 @@ import LineGraph from 'shared/components/LineGraph';
 import SingleStat from 'shared/components/SingleStat';
 import NameableGraph from 'shared/components/NameableGraph';
 import ReactGridLayout, {WidthProvider} from 'react-grid-layout';
+
+import timeRanges from 'hson!../data/timeRanges.hson';
+
 const GridLayout = WidthProvider(ReactGridLayout);
 
 const RefreshingLineGraph = AutoRefresh(LineGraph);
@@ -21,23 +24,15 @@ export const LayoutRenderer = React.createClass({
   propTypes: {
     autoRefresh: number.isRequired,
     timeRange: shape({
-      defaultGroupBy: string.isRequired,
-      queryValue: string.isRequired,
+      lower: string.isRequired,
     }).isRequired,
     cells: arrayOf(
       shape({
         queries: arrayOf(
           shape({
             label: string,
-            range: shape({
-              upper: number,
-              lower: number,
-            }),
-            rp: string,
-            text: string.isRequired,
-            database: string.isRequired,
-            groupbys: arrayOf(string),
-            wheres: arrayOf(string),
+            text: string,
+            query: string,
           }).isRequired
         ).isRequired,
         x: number.isRequired,
@@ -46,6 +41,7 @@ export const LayoutRenderer = React.createClass({
         h: number.isRequired,
         i: string.isRequired,
         name: string.isRequired,
+        type: string.isRequired,
       }).isRequired
     ),
     host: string,
@@ -54,15 +50,17 @@ export const LayoutRenderer = React.createClass({
     onEditCell: func,
     onRenameCell: func,
     onUpdateCell: func,
+    onSummonOverlayTechnologies: func,
   },
 
   buildQuery(q) {
-    const {timeRange, host} = this.props;
-    const {wheres, groupbys} = q;
+    const {timeRange: {lower}, host} = this.props
+    const {defaultGroupBy} = timeRanges.find((range) => range.lower === lower)
+    const {wheres, groupbys} = q
 
     let text = q.text;
 
-    text += ` where time > ${timeRange.queryValue}`;
+    text += ` where time > ${lower}`;
 
     if (host) {
       text += ` and \"host\" = '${host}'`;
@@ -76,25 +74,25 @@ export const LayoutRenderer = React.createClass({
       if (groupbys.find((g) => g.includes("time"))) {
         text += ` group by ${groupbys.join(',')}`;
       } else if (groupbys.length > 0) {
-        text += ` group by time(${timeRange.defaultGroupBy}),${groupbys.join(',')}`;
+        text += ` group by time(${defaultGroupBy}),${groupbys.join(',')}`;
       } else {
-        text += ` group by time(${timeRange.defaultGroupBy})`;
+        text += ` group by time(${defaultGroupBy})`;
       }
     } else {
-      text += ` group by time(${timeRange.defaultGroupBy})`;
+      text += ` group by time(${defaultGroupBy})`;
     }
 
     return text;
   },
 
   generateVisualizations() {
-    const {autoRefresh, source, cells, onEditCell, onRenameCell, onUpdateCell} = this.props;
+    const {autoRefresh, source, cells, onEditCell, onRenameCell, onUpdateCell, onSummonOverlayTechnologies} = this.props;
 
     return cells.map((cell) => {
-      const qs = cell.queries.map((q) => {
-        return Object.assign({}, q, {
+      const qs = cell.queries.map((query) => {
+        return Object.assign({}, query, {
           host: source,
-          text: this.buildQuery(q),
+          text: this.buildQuery(query),
         });
       });
 
@@ -105,6 +103,7 @@ export const LayoutRenderer = React.createClass({
               onEditCell={onEditCell}
               onRenameCell={onRenameCell}
               onUpdateCell={onUpdateCell}
+              onSummonOverlayTechnologies={onSummonOverlayTechnologies}
               cell={cell}
             >
               <RefreshingSingleStat queries={[qs[0]]} autoRefresh={autoRefresh} />
@@ -124,12 +123,13 @@ export const LayoutRenderer = React.createClass({
             onEditCell={onEditCell}
             onRenameCell={onRenameCell}
             onUpdateCell={onUpdateCell}
+            onSummonOverlayTechnologies={onSummonOverlayTechnologies}
             cell={cell}
           >
             <RefreshingLineGraph
               queries={qs}
               autoRefresh={autoRefresh}
-              showSingleStat={cell.type === "line-plus-single-stat"}
+              showSingleStat={cell.type === 'line-plus-single-stat'}
               displayOptions={displayOptions}
             />
           </NameableGraph>