+++ ui/src/shared/components/NameableGraph.js
@@ -1,60 +1,9 @@
 import React, {PropTypes} from 'react'
-
-const NameableGraph = ({
-  cell: {
-    name,
-    isEditing,
-    x,
-    y,
-  },
-  cell,
-  children,
-  onEditCell,
-  onRenameCell,
-  onUpdateCell,
-}) => {
-  let nameOrField
-  const isEditable = !!(onEditCell || onRenameCell || onUpdateCell)
-
-  if (isEditing && isEditable) {
-    nameOrField = (
-      <input
-        type="text"
-        value={name}
-        autoFocus={true}
-        onChange={onRenameCell(x, y)}
-        onBlur={onUpdateCell(cell)}
-        onKeyUp={(evt) => {
-          if (evt.key === 'Enter') {
-            onUpdateCell(cell)()
-          }
-        }}
-      />
-    )
-  } else {
-    nameOrField = name
-  }
-
-  let onClickHandler
-  if (isEditable) {
-    onClickHandler = onEditCell
-  } else {
-    onClickHandler = () => {
-      // no-op
-    }
-  }
-
-  return (
-    <div>
-      <h2 className="dash-graph--heading" onClick={onClickHandler(x, y, isEditing)}>{nameOrField}</h2>
-      <div className="dash-graph--container">
-        {children}
-      </div>
-    </div>
-  )
-}
+import classnames from 'classnames'
+import OnClickOutside from 'react-onclickoutside'
 
 const {
+  bool,
   func,
   node,
   number,
@@ -62,16 +11,109 @@ const {
   string,
 } = PropTypes
 
-NameableGraph.propTypes = {
-  cell: shape({
-    name: string.isRequired,
-    x: number.isRequired,
-    y: number.isRequired,
-  }).isRequired,
-  children: node.isRequired,
-  onEditCell: func,
-  onRenameCell: func,
-  onUpdateCell: func,
-}
+const NameableGraph = React.createClass({
+  propTypes: {
+    cell: shape({
+      name: string.isRequired,
+      isEditing: bool,
+      x: number.isRequired,
+      y: number.isRequired,
+    }).isRequired,
+    children: node.isRequired,
+    onEditCell: func,
+    onRenameCell: func,
+    onUpdateCell: func,
+    onSummonOverlayTechnologies: func,
+  },
 
+  getInitialState() {
+    return {
+      isMenuOpen: false,
+    }
+  },
+
+  toggleMenu() {
+    this.setState({
+      isMenuOpen: !this.state.isMenuOpen,
+    })
+  },
+
+  closeMenu() {
+    this.setState({
+      isMenuOpen: false,
+    })
+  },
+
+  render() {
+    const {
+      cell,
+      cell: {
+        x,
+        y,
+        name,
+        isEditing,
+      },
+      onEditCell,
+      onRenameCell,
+      onUpdateCell,
+      onSummonOverlayTechnologies,
+      children,
+    } = this.props
+
+    const isEditable = !!(onEditCell || onRenameCell || onUpdateCell)
+
+    let nameOrField
+    if (isEditing && isEditable) {
+      nameOrField = (
+        <input
+          className="form-control input-sm dash-graph--name-edit"
+          type="text"
+          value={name}
+          autoFocus={true}
+          onChange={onRenameCell(x, y)}
+          onBlur={onUpdateCell(cell)}
+          onKeyUp={(evt) => {
+            if (evt.key === 'Enter') {
+              onUpdateCell(cell)()
+            }
+          }}
+        />
+      )
+    } else {
+      nameOrField = (<span className="dash-graph--name">{name}</span>)
+    }
+
+    let onClickHandler
+    if (isEditable) {
+      onClickHandler = onEditCell
+    } else {
+      onClickHandler = () => {
+        // no-op
+      }
+    }
+
+    return (
+      <div className="dash-graph">
+        <div className="dash-graph--heading">
+          <div onClick={onClickHandler(x, y, isEditing)}>{nameOrField}</div>
+          <ContextMenu isOpen={this.state.isMenuOpen} toggleMenu={this.toggleMenu} onSummonOverlayTechnologies={onSummonOverlayTechnologies} cell={cell} handleClickOutside={this.closeMenu}/>
+        </div>
+        <div className="dash-graph--container">
+          {children}
+        </div>
+      </div>
+    )
+  },
+})
+
+const ContextMenu = OnClickOutside(({isOpen, toggleMenu, onSummonOverlayTechnologies, cell}) => (
+  <div className={classnames("dash-graph--options", {"dash-graph--options-show": isOpen})} onClick={toggleMenu}>
+    <button className="btn btn-info btn-xs">
+      <span className="icon caret-down"></span>
+    </button>
+    <ul className="dash-graph--options-menu">
+      <li onClick={() => onSummonOverlayTechnologies(cell)}>Edit</li>
+    </ul>
+  </div>
+))
 export default NameableGraph;