+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
   1. [#1020](https://github.com/influxdata/chronograf/pull/1020): Users can now edit cell names on dashboards
+  2. [#1035](https://github.com/influxdata/chronograf/pull/1035): Convert many InfluxQL statements to query builder
+  3. [#1015](https://github.com/influxdata/chronograf/pull/1015): Introduce ability to edit a dashboard cell
 
 ### UI Improvements